using `GroupedNaturalIdPartitioner`; simplifying `FilteredLeRobotDataset`; update authors; `README.md` teaks
