+++ examples/quickstart-lerobot/lerobot_example/lerobot_federated_dataset.py
@@ -1,50 +1,53 @@
-# Copyright 2023 Flower Labs GmbH. All Rights Reserved.
-# Copyright 2024 zk0 DBA. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""FederatedLeRobotDataset."""
+from pathlib import Path
+from typing import Callable
 
+from lerobot.common.datasets.lerobot_dataset import (
+    CODEBASE_VERSION,
+    DATA_DIR,
+    LeRobotDataset,
+)
+from lerobot.common.datasets.utils import (
+    calculate_episode_data_index,
+    load_info,
+    load_stats,
+    load_videos,
+    reset_episode_index,
+)
 
-from flwr_datasets import FederatedDataset
+from datasets import Dataset
 
+class FilteredLeRobotDataset(LeRobotDataset):
+    """Delays loading and processing of dataset until load function is called with an
+    optional filter argument."""
 
-class FederatedLeRobotDataset(FederatedDataset):
-    def _prepare_dataset(self) -> None:
-        """Overrides the implementation of FederatedDataset in order to account for
-        LeRobot specific dataset loading.
+    def __init__(
+        self,
+        repo_id: str,
+        hf_dataset: Dataset,
+        root: Path | None = DATA_DIR,
+        split: str = "train",
+        image_transforms: Callable | None = None,
+        delta_timestamps: dict[list[float]] | None = None,
+        video_backend: str | None = None,
+    ):
+        self.repo_id = repo_id
+        self.root = root
+        self.split = split
+        self.image_transforms = image_transforms
+        self.delta_timestamps = delta_timestamps
+        self.video_backend = video_backend
+        self.hf_dataset = hf_dataset
 
-        Prepare the dataset (prior to partitioning) by downloading.
+        # after filtering, the stored episode data index may not be the same
+        # so let's calculate it on the filtered data
+        self.episode_data_index = calculate_episode_data_index(self.hf_dataset)
+        self.hf_dataset = reset_episode_index(self.hf_dataset)
 
-        Currently only "train" dataset is loaded from LeRobot datasets. Test datasets are generated via gym.
+        self.stats = load_stats(self.repo_id, CODEBASE_VERSION, self.root)
+        self.info = load_info(self.repo_id, CODEBASE_VERSION, self.root)
+        if self.video:
+            self.videos_dir = load_videos(self.repo_id, CODEBASE_VERSION, self.root)
+            self.video_backend = (
+                self.video_backend if self.video_backend is not None else "pyav"
+            )
 
-        Run only ONCE when triggered by load_* function. (In future more control whether
-        this should happen lazily or not can be added). The operations done here should
-        not happen more than once.
-
-        It is controlled by a single flag, `_dataset_prepared` that is set True at the
-        end of the function.
-
-        NOTE:
-        Currently the following two parameters from FederatedDataset are not supported:
-        subset - dataset subset loading not supported.
-        shuffle - dataset shuffling not supported in this type of dataset.
-        """
-        self._dataset = {
-            "train": {"dataset_name": self._dataset_name, **self._load_dataset_kwargs}
-        }
-        if self._preprocessor:
-            self._dataset = self._preprocessor(self._dataset)
-        available_splits = list(self._dataset.keys())
-        self._event["load_split"] = {split: False for split in available_splits}
-        self._dataset_prepared = True
